put reservation under resourceRoute in backend
