+++ server/src/routes/resourceRoute.ts
@@ -1,5 +1,5 @@
 import { Router } from "express";
-
+import {verifyToken} from "../middleware/middleware";
 import controller from "../controllers/resourceController";
 
 const resourceRouter = Router();
@@ -18,5 +18,7 @@ resourceRouter.get("/languages", controller.queryResourceLanguages);
 
 // get resources by _id
 resourceRouter.get("/:id", controller.queryResourceById);
+
+resourceRouter.post('/reservation', verifyToken, controller.postReservation)
 
 export default resourceRouter;
